requirements.txt
openai
streamlit
youtube-transcript-api
python-dotenv
